dsp16xx.c (print_operand): Fix format specifier.
	* dsp16xx.c (print_operand): Fix format specifier.
	* dsp16xx.md: Avoid automatic aggregate initialization.
	* frv.h (REG_CLASS_FROM_LETTER): Avoid char as array index.
	* h8300.c (emit_a_rotate, h8300_adjust_insn_length): Avoid U
	integer constant modifier.
	* ip2k.c (ip2k_set_compare): Avoid signed/unsigned warning.
	* mmix-protos.h (mmix_use_simple_return): Move outside TREE_CODE
	guards.
	* sh/netbsd-elf.h (FUNCTION_PROFILER): Fix format specifier.
	* v850.c (v850_select_section): Mark parameter with
	ATTRIBUTE_UNUSED.
	* global.c (global_alloc): Const-ify.
	* ra-colorize.c (hardregset_to_string): Fix format specifier.

From-SVN: r56212

+++ gcc/config/h8300/h8300.c
@@ -3212,7 +3212,7 @@ emit_a_rotate (code, operands)
 
   /* Determine the faster direction.  After this phase, amount will be
      at most a half of GET_MODE_BITSIZE (mode).  */
-  if ((unsigned int) amount > GET_MODE_BITSIZE (mode) / 2U)
+  if ((unsigned int) amount > GET_MODE_BITSIZE (mode) / (unsigned) 2)
     {
       /* Flip the direction.  */
       amount = GET_MODE_BITSIZE (mode) - amount;
@@ -3747,7 +3747,7 @@ h8300_adjust_insn_length (insn, length)
 
       /* Determine the faster direction.  After this phase, amount
 	 will be at most a half of GET_MODE_BITSIZE (mode).  */
-      if ((unsigned int) amount > GET_MODE_BITSIZE (mode) / 2U)
+      if ((unsigned int) amount > GET_MODE_BITSIZE (mode) / (unsigned) 2)
 	/* Flip the direction.  */
 	amount = GET_MODE_BITSIZE (mode) - amount;
